aula 3

+++ src/main/java/br/mba/fiap/abctech/abctechservice/model/OrderLocation.java
@@ -6,7 +6,8 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 import javax.persistence.*;
-import java.sql.Date;
+import java.util.Date;
+
 
 @Entity
 @Getter
@@ -19,8 +20,8 @@ public class OrderLocation {
     @Column(name = "id", nullable = false)
     private Long id;
 
-    private Integer latitude;
-    private Integer longitude;
+    private Double latitude;
+    private Double longitude;
     private Date date;
 }
